adding conversion functionality
now the tool fixes out teams when they represent >x% of the N best paths. This results in the tool eventually converging to a local maximum. Future versions will work on reducing the number of paths tried and focusing on likely to be good paths to improve efficiency and allow broader initial settings to make it more likely to reach a global maximum